added javascript import for animation component
Backwards compatibility is preserved with this change.

This extra script could have just been added to index.html,
but for projects that didn't use the animation component, that would
bloat the output file unnecessarily.

The included script should probably be moved to a ku-fpg location for
safety, but I don't know exactly where to put that or how to do so.